fix: remove unnecessary null check for parser_rules in dynamic categories list fix: remove unnecessary null check for parser_rules in dynamic categories list fix: update androidx test libraries to latest versions chore: update gradle wrapper to version 9.5.1

+++ app/src/main/java/com/example/islandlyrics/ui/common/LayeredPagerTransition.kt
@@ -52,7 +52,7 @@ fun LayeredPagerPage(
             modifier = Modifier
                 .fillMaxSize()
                 .graphicsLayer {
-                    val width = size.width.toFloat()
+                    val width = size.width
                     if (isIncoming) {
                         translationX = -clampedOffset * width
                         val scale = 1.01f - (0.01f * incomingProgress)
